pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = ""
readme = "README.md"
authors = []
packages = [
    { include = "back" },
    { include = "tests" },
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
sqlmodel = "^0.0.16"
bcrypt = "^4.1.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.7"
pytest = "^8.1.1"
httpx = "^0.27.0"
pytest-xdist = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "back.main:run_dev_server"